Add showing running server Add log font settings Add update checking

+++ KcptunLauncher/Controller/MainProcessController.cs
@@ -10,38 +10,29 @@ namespace KcptunLauncher.Controller
 {
     public class MainProcessController
     {
-        public static string KCPTUN_CLIENT_FILE_PATH = AppDomain.CurrentDomain.BaseDirectory + "client_windows_amd64.exe";
+        public static string KcptunClientFilePath = AppDomain.CurrentDomain.BaseDirectory + "client_windows_amd64.exe";
 
-        private static MainProcessController _Instance;
+        private static MainProcessController _instance;
         public static MainProcessController GetInstance()
         {
-            if (_Instance == null) { _Instance = new MainProcessController(); }
-            return _Instance;
+            return _instance ?? (_instance = new MainProcessController());
         }
 
-        private Hashtable mProcesses;
-        public Hashtable Processes { get { return mProcesses; } }
+        public Hashtable Processes { get; }
 
         private MainProcessController()
         {
-            mProcesses = new Hashtable();
+            Processes = new Hashtable();
         }
 
         public bool CheckKcptunExists()
         {
-            if (File.Exists(KCPTUN_CLIENT_FILE_PATH))
-            {
-                return true;
-            }
-            else
-            {
-                return false;
-            }
+            return File.Exists(KcptunClientFilePath);
         }
 
         public bool CheckKcptunRunning(string serverName)
         {
-            return mProcesses.ContainsKey(serverName);
+            return Processes.ContainsKey(serverName);
         }
 
         public EventHandler ProcessLogReceivedHandler;
@@ -60,12 +51,12 @@ namespace KcptunLauncher.Controller
 
         public bool Start(Server server)
         {
-            if (mProcesses.ContainsKey(server.Name))
+            if (Processes.ContainsKey(server.Name))
             {
                 return true;
             }
 
-            if (!File.Exists(KCPTUN_CLIENT_FILE_PATH))
+            if (!File.Exists(KcptunClientFilePath))
             {
                 if (System.Windows.Forms.MessageBox.Show("未找到Kcptun客户端，是否查找？", "kcptun", System.Windows.Forms.MessageBoxButtons.YesNo, System.Windows.Forms.MessageBoxIcon.Question) == System.Windows.Forms.DialogResult.Yes)
                 {
@@ -76,7 +67,7 @@ namespace KcptunLauncher.Controller
             }
 
             Process mProcess = new Process();
-            mProcess.StartInfo.FileName = KCPTUN_CLIENT_FILE_PATH;
+            mProcess.StartInfo.FileName = KcptunClientFilePath;
             mProcess.StartInfo.Arguments = server.ToString();
 
             mProcess.StartInfo.UseShellExecute = false;
@@ -85,10 +76,10 @@ namespace KcptunLauncher.Controller
             mProcess.StartInfo.RedirectStandardInput = true;
             mProcess.StartInfo.RedirectStandardOutput = true;
 
-            mProcess.ErrorDataReceived += new DataReceivedEventHandler((object sender, DataReceivedEventArgs e) => { ProcessLogReceived(mProcess, server, e.Data); });
-            mProcess.OutputDataReceived += new DataReceivedEventHandler((object sender, DataReceivedEventArgs e) => { ProcessLogReceived(mProcess, server, e.Data); });
+            mProcess.ErrorDataReceived += (sender, e) => { ProcessLogReceived(mProcess, server, e.Data); };
+            mProcess.OutputDataReceived += (sender, e) => { ProcessLogReceived(mProcess, server, e.Data); };
 
-            mProcesses.Add(server.Name, mProcess);
+            Processes.Add(server.Name, mProcess);
 
             mProcess.Start();
             mProcess.BeginErrorReadLine();
@@ -110,22 +101,21 @@ namespace KcptunLauncher.Controller
 
         public void Stop(string serverName)
         {
-            if (mProcesses.ContainsKey(serverName))
+            if (!Processes.ContainsKey(serverName)) return;
+
+            if (Configuration.EnabledServerList.Contains(serverName))
             {
-                if (Configuration.EnabledServerList.Contains(serverName))
-                {
-                    Configuration.EnabledServerList.Remove(serverName);
-                }
-                Process mProcess = mProcesses[serverName] as Process;
-                if (!mProcess.HasExited)
-                {
-                    mProcess.Kill();
-                    mProcess.CancelErrorRead();
-                    mProcess.CancelOutputRead();
-                    mProcess = null;
-                }
-                mProcesses.Remove(serverName);
+                Configuration.EnabledServerList.Remove(serverName);
             }
+
+            Process mProcess = Processes[serverName] as Process;
+            if (mProcess != null && !mProcess.HasExited)
+            {
+                mProcess.Kill();
+                mProcess.CancelErrorRead();
+                mProcess.CancelOutputRead();
+            }
+            Processes.Remove(serverName);
         }
 
         public void StopAll()
